MOM-249 Organize packages
