+++ src/main/java/no/dusken/momus/service/remotedocument/RemoteDocument.java
@@ -1,0 +1,6 @@
+package no.dusken.momus.service.remotedocument;
+
+public interface RemoteDocument {
+    String getId();
+    
+}
