+++ src/main/java/no/dusken/momus/service/remotedocument/sharepoint/models/Drive.java
@@ -1,18 +1,16 @@
-package no.dusken.momus.service.sharepoint.models;
+package no.dusken.momus.service.remotedocument.sharepoint.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 
 @Getter
 @Setter
 @NoArgsConstructor
-@ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class User {
+public class Drive {
     private String id;
-    private String displayName;
+    private String driveType;
 }
