+++ src/main/java/no/dusken/momus/service/remotedocument/sharepoint/models/DriveItem.java
@@ -1,4 +1,4 @@
-package no.dusken.momus.service.sharepoint.models;
+package no.dusken.momus.service.remotedocument.sharepoint.models;
 
 import java.time.LocalDate;
 
@@ -8,13 +8,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
+import no.dusken.momus.service.remotedocument.RemoteDocument;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class DriveItem {
+public class DriveItem implements RemoteDocument {
     private String id;
     private String name;
     private String webUrl;
