+++ src/main/java/no/dusken/momus/service/remotedocument/sharepoint/models/DriveItemChildren.java
@@ -1,17 +1,17 @@
-package no.dusken.momus.service.sharepoint.models;
+package no.dusken.momus.service.remotedocument.sharepoint.models;
+
+import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 
 @Getter
 @Setter
 @NoArgsConstructor
-@ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class FolderResource {
-    private int childCount;
+public class DriveItemChildren {
+    private List<DriveItem> value;
 }
