+++ src/main/java/no/dusken/momus/service/remotedocument/sharepoint/models/Entity.java
@@ -1,4 +1,4 @@
-package no.dusken.momus.service.sharepoint.models;
+package no.dusken.momus.service.remotedocument.sharepoint.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
@@ -12,6 +12,6 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class FileResource {
-    private String mimeType;
+public class Entity {
+    private User user;
 }
