+++ src/main/java/no/dusken/momus/service/remotedocument/sharepoint/models/FileResource.java
@@ -1,17 +1,17 @@
-package no.dusken.momus.service.sharepoint.models;
-
-import java.util.List;
+package no.dusken.momus.service.remotedocument.sharepoint.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.ToString;
 
 @Getter
 @Setter
 @NoArgsConstructor
+@ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class DriveItemChildren {
-    private List<DriveItem> value;
+public class FileResource {
+    private String mimeType;
 }
