+++ src/main/java/no/dusken/momus/service/remotedocument/sharepoint/models/FolderResource.java
@@ -1,16 +1,17 @@
-package no.dusken.momus.service.sharepoint.models;
+package no.dusken.momus.service.remotedocument.sharepoint.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.ToString;
 
 @Getter
 @Setter
 @NoArgsConstructor
+@ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Drive {
-    private String id;
-    private String driveType;
+public class FolderResource {
+    private int childCount;
 }
